newest stuff!

+++ libc/string.h
@@ -7,8 +7,8 @@ extern "C" {
 
 #include <stdlib.h>
 
-void *_memsetd(void *s, unsigned int f, unsigned int n);
-void *memset(void *s, int c, unsigned int n);
+void *_memsetd(void *s, unsigned int f, size_t n);
+void *memset(void *s, int c, size_t n);
 void *memcpy(void *, const void *, size_t);
 char *strncpy(char *dest, const char *src, size_t n);
 void *memmove(void *, const void *, size_t);
